Task done
This task implements a random string generator. The user is prompted to select the length of the string (minimum length 0, maximum length 20) and content (numbers, uppercase and lowercase letters).

After pressing the "Generate" button, a check is made for the length of the string so that it is not equal to zero and that at least one option is selected.

After a successful check, the generated string is displayed in the "Result" field.